packages/server/requirements.txt
flask
geojson
geopandas
psycopg2-binary
python-dotenv
sqlalchemy
